Ignore downloaded Ansible Galaxy roles
Support having Ansible download the Galaxy roles into the working copy
of the repository instead of globally across the developer machine’s
host. Updates the .gitignore file to ignore both the downloaded Ansible
Galaxy files and a repository-specific ansible.cfg. The README is
updated to document the syntax needed for the custom ansible.cfg.